Renamed "dataHandeler" to "dataHandler" Fixed a few other mispells here and there.
