+++ core/src/main/java/wtf/nucker/kitpvpplus/dataHandlers/PlayerData.java
@@ -1,4 +1,4 @@
-package wtf.nucker.kitpvpplus.dataHandelers;
+package wtf.nucker.kitpvpplus.dataHandlers;
 
 import org.bukkit.OfflinePlayer;
 import wtf.nucker.kitpvpplus.objects.Kit;
