+++ core/src/main/java/wtf/nucker/kitpvpplus/dataHandlers/PlayerState.java
@@ -1,4 +1,4 @@
-package wtf.nucker.kitpvpplus.dataHandelers;
+package wtf.nucker.kitpvpplus.dataHandlers;
 
 /**
  * @author Nucker
