+++ core/src/main/java/wtf/nucker/kitpvpplus/integrations/worldguard/WorldGuardLegacy.java
@@ -11,8 +11,8 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerMoveEvent;
 import wtf.nucker.kitpvpplus.KitPvPPlus;
-import wtf.nucker.kitpvpplus.dataHandelers.PlayerData;
-import wtf.nucker.kitpvpplus.dataHandelers.PlayerState;
+import wtf.nucker.kitpvpplus.dataHandlers.PlayerData;
+import wtf.nucker.kitpvpplus.dataHandlers.PlayerState;
 import wtf.nucker.kitpvpplus.flags.ArenaFlag;
 import wtf.nucker.kitpvpplus.flags.ProtectedFlag;
 import wtf.nucker.kitpvpplus.flags.SpawnFlag;
